Reverts fix on JoinableFile for commit 7b29a31
  After checking, the JoinableFile.joinStream should still be in
synchronized block. The usage of the unblocking way is not reasonable
because it should import some dirty read problems for the joinable and
jointCount.
  For example, when Thread 1(T1) is using reallyClose(), it sets the
joinable = false, but not finished the left work. If the joinStream is
not synchronized, and the Thread 2(T2) will start reading the file,
it can find this file becausethe reallyClose is not done yet,
but found that the joinable is false which may not let this joinStream
operation success. I think this is weird and not correct.