Improve concurrency when callers race to be first to open read stream.
Multiple callers trying to be the first to open a new joinable stream
were resulting in OverlappingFileLockException because the FileEntry
'file' field wasn't sufficiently protected.

New code adds extra synchronization and a tryLock() approach to locking
a re-entrant operation, with while-looping and a timeout to locking
attempts. This will hopefully guard against this problem without incurring
too much of a performance hit.