Change operationLocks map to ConcurrentHashMap. Use blocking lock() call for opLocks
I've switched to the blocking form of the ReentrantLock.lock() call to avoid the need
to loop and timeout, which adds a lot of complexity. I need to finish cleaning this up
so this isn't the last commit for the PR.

I also switched the map type for operation locks to a ConcurrentHashMap, per @ruhan1's
suggestion. This may make memory usage drift upward over time, so we should investigate
embedding this in the FileEntry or something to control the lifecycle of the ReentrantLock.

More to come, but this at least will eliminate the OverlappingFileLockException.