Add documentation, and remove operationLock when clearLocks cleans up
Adding documentation for most things (major things) to try to make it
clearer how the different types of locks interact, and why they're
necessary.

Also, add cleanup step to FileTree.clearLocks() to remove the FileOperationLock
from the operationLocks map at the same time the FileEntry is removed
from the entries map. This should hopefully keep a lid on memory consumption.