Remove code that cleans up operationLocks on file close.
This change brought back the dreaded OverlappingFileLockException.

Worse, we don't have tests in this project to reliably detect that
problem...I only found it through testing of Indy under loads designed
to produce collisions on specific artifacts.

I think what was happening with this cleanup measure is that the clearLocks()
method (called from JoinableFile.close()) was removing the ReentrantLock
too quickly, when normally the method would complete and unlock the
ReentrantLock for the next caller. In this case though, removing that lock
allowed the next operation to create a new one and lock it, thinking it
was the only one accessing the file (or maybe just first). Looking at
the JoinableFile.reallyClose() method, it looks like the extra delay
normally taken to unlock that ReentrantLock is just enough to allow
the FileLock to clear (the lock is explicitly released a few lines before
the callback would trigger clearLocks).

This suggests that it might still be possible to cause OverlappingFileLockException
if the lock was even slower to release. It's not clear how to determine
whether the lock was actually released (before proceeding).

In any case, I'm rolling back this cleanup measure because it provides
enough stability to resist moderate load (at least).