Disable use of FileLock
From my reading of the docs about FileLock, it seems like it may not be safe
for use with RandomAccessFile + FileChannel, particularly when we're mapping
specific ranges of bytes into buffers for different readers. This isn't
necessary for internal consistency anyway, it's mainly to prevent external
processes from modifying the files that partyline is managing. This is a
fairly frige concern, I think.

Simply commenting FileLock + usages for now.