Add ownership lock counting to allow multiple directory lock operations
We need a set of coordinated threads to be able to lock a directory repeatedly
without the need to figure out which thread locked it (and which one should
unlock it).

This change adds a LockOwnerInfo inner class which enables tracking of
a lock count per owner. It also removes the lock ownerName from explicit
specification, using ThreadContext to store a single owner and manage locks
at the context level.