Process associated dir entries as a linked list for this file when locking/unlocking.
When we attempt to lock or unlock a file (or even find out if is locked currently), we
consider locks in the surrounding filesystem. Locks on child files imply a read lock on
directory ancestry. Locks on ancestor directories limit the locks available for descendant
files. When we place a lock on a file successfully, that should increase the read-lock
count for any already-locked ancestor directories...this will prevent directory locks from
becoming decoupled from the conditions under which the file lock was established.

This means that any pre-existing directory lock related to our file become linked into the
FileEntry for our file...or our directory, if that's what it is. Over successive operations
this will create a linked-list data structure composed of pre-existing locked FileEntries
in the directory structure of our file. When we lock or unlock our file, then, we need to
adjust the lock counts in these linked entries. It's not enough to process only the parent
entry...we have to adjust the whole list.

This change builds on @ligangty's work (and some of my own) to improve accounting for
directory locking. It processes the full linked list of associated directory FileEntry
instances for any given file, at the point where that file's locks are adjusted. When a
file is locked, its ancestry with pre-existing FileEntries get a new read lock. When
a file is unlocked, its associated ancestor FileEntries get unlocked (for this context,
and just to decrement the lock count...never to clear the locks blindly).

With this change, all existing tests work (except those previously @Ignored for other
reasons).