Improve ThreadContext synchronization, use ThreadContext finalizer (new), and fix JoinableInputStream hashcode
JoinableInputStream hashcode appeared wrong (using the default one from Object), which seemed to be
leading to problems in jointClosed() when removing the stream from the list of open streams. To fix,
I added a hashCode method that uses the original thread name, joint index, and a new ctorTime field,
which uses System.nanoTime(). All three of these would have to collide to produce a duplicate, which is
highly unlikely.

ThreadContext and cleanupCurrentThread() had a weakness where the "topmost" thread in a request could
call cleanupCurrentThread() before child threads were finished, which would have the effect of closing
streams and removing the open streams map from the context, all before the child thread was done with
them. To fix this, I introduced finalizer support to ThreadContext (weft 1.5) and then moved the
cleanupCurrentThread() logic into a finalizer, which gets registered in the ThreadContext when a new
stream is added via the JoinableFileManager.

Finally, I added synchronization blocks around ThreadContext operations, keyed to the ThreadContext
instance, to prevent weirdness around setting/removing state in the context.