turn off stream finalizer registration
We're having problems in Indy where certain streams are opened inside a ThreadContext
and get registered, then are handed off to threads that are outside the context for
streaming back to the user (and there's another more mysterious case happening when
trying to merge metadata files in Indy).

If we have good coding discipline inside Indy, we shouldn't need this finalizer to
come along behind and clean up the streams...they should already be cleaned up as we go.

This change simply disables registration of streams in the ThreadContext finalizer.